now adds guidelines

+++ src/pandora.sh
@@ -52,7 +52,8 @@ then
         --clinvar_api_key /home/dnanexus/in/clinvar_api_key/*.txt \
         --clinvar_testing $clinvar_testing
     mkdir -p /home/dnanexus/out/clinvar_accession_id
-    mv accession_ids.txt /home/dnanexus/out/clinvar_accession_id
+    # If file exists move to output directory
+    [ -f accession_ids.txt ] && mv accession_ids.txt /home/dnanexus/out/clinvar_accession_id
     dx-upload-all-outputs
 else
     echo Running mode $running_mode is not valid please choose one of the following:
